% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.to.dictionary.outcome.r
\name{add.to.dictionary.outcome}
\alias{add.to.dictionary.outcome}
\title{Worker function that fits outcome models}
\usage{
add.to.dictionary.outcome(Y, X, U, W, alpha, binary = F)
}
\arguments{
\item{Y}{vector of the outcome}

\item{X}{vector of the treatment (0/1)}

\item{U}{matrix of covariates to be considered for inclusion/exclusion}

\item{W}{matrix of covariates that will be included in all models (optional)}

\item{alpha}{vector of inclusion indicators (which columns of U) to included in the propensity score model}

\item{binary}{indicates if the outcome is binary}
}
\value{
A list. The list contains the following named components:
\item{out}{a list that contains the BIC, predicted values, and estimated treatment effect from each outcome model}
}
\description{
This function fits outcome models and saves necessary information
}

