% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.to.dictionary.r
\name{add.to.dictionary}
\alias{add.to.dictionary}
\title{Worker function that fits propensity score models}
\usage{
add.to.dictionary(X, U, W, alpha)
}
\arguments{
\item{X}{vector of the treatment (0/1)}

\item{U}{matrix of covariates to be considered for inclusion/exclusion}

\item{W}{matrix of covariates that will be included in all models (optional)}

\item{alpha}{vector of inclusion indicators (which columns of U) to included in the propensity score model}
}
\value{
A list. The list contains the following named components:
\item{out}{a list that contains the BIC and estimated propensity scores from propensity score models}
}
\description{
This function fits propensity score models and saves necessary information
}

