% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.to.prob.r
\name{bic.to.prob}
\alias{bic.to.prob}
\title{Convert BIC to model probabilities}
\usage{
bic.to.prob(bic)
}
\arguments{
\item{bic}{vector of BICs}
}
\value{
A vector of model probabilities of the same dimension of \code{bic}
}
\description{
This function transforms BIC to model probabilities
}

