% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLikelihood.R
\name{calcLikelihood}
\alias{calcLikelihood}
\title{Calculate the likelihood for a MADproject object.}
\usage{
calcLikelihood(proj, data)
}
\arguments{
\item{proj}{The MADproject object with data read from the MAD# databases.}

\item{data}{The kind of data to use for the likelihood calculations.}
}
\value{
likes A list of matching length as \code{data} the likelihood values 
  for each sample based on \code{data}.
}
\description{
\code{calcLikelihood} returns the likelihood values based on the 
observation and realization data in MADproject and the requested
inversion data type \code{data}.
}

