% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPosterior.R
\name{calcPosterior}
\alias{calcPosterior}
\title{Calculate the posterior for a \code{MADproject} object.}
\usage{
calcPosterior(proj, likes)
}
\arguments{
\item{proj}{The \code{MADproject} object with the \code{likelihood}
slot filled.}
}
\value{
proj An updated \code{MADproject} object with the
\code{posterior} slot filled.
}
\description{
\code{calcPosterior} returns the posterior values for each sample
}

