% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.r
\name{expit}
\alias{expit}
\title{Expit (inverse logit) function}
\usage{
expit(x)
}
\arguments{
\item{x}{vector of values to apply the expit function}
}
\value{
A vector of the same dimension of \code{x}
}
\description{
This function transforms the input using the expit function
}

