% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADproject.R
\docType{class}
\name{MADproject-class}
\alias{MADproject}
\alias{MADproject-class}
\title{An S4 class to represent a MAD project}
\description{
An S4 class to represent a MAD project
}
\section{Slots}{

\describe{
\item{\code{numLocations}}{A length-one numeric vector for the number of measurement locations}

\item{\code{numTimesteps}}{A length-one numeric vector for the number of time steps in series}

\item{\code{numSamples}}{A length-one numeric vector for the number of samples}

\item{\code{observations}}{A numeric \code{numTimesteps}-by-\code{numLocations} matrix
for the observations at the m}

\item{\code{realizations}}{A list of length \code{numSamples} with each element containing
a numeric (number of realizations)-by-\code{numLocations*numTimesteps} matrix
to hold the ensemble of simulated measurements.}
}}

