% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madr.r
\name{madr}
\alias{madr}
\title{Calculate model averaged double robust estimate}
\usage{
madr(Y, X, U, W = NULL, M = 1000, cut = 0.95, enumerate = F,
  tau = NULL, two.stage = NULL)
}
\arguments{
\item{Y}{vector of the outcome}

\item{X}{vector of the treatment (0/1)}

\item{U}{matrix of covariates to be considered for inclusion/exclusion}

\item{W}{matrix of covariates that will be included in all models (optional)}

\item{M}{the number of MCMC iteration}

\item{cut}{cumulative probability of models to be retained for improved computational efficiency (1 retains all visited models)}

\item{enumerate}{indicator if all possible models should be enumerated (default: FALSE)}

\item{tau}{scalar value for the prior model dependence (1 is an independent prior; defaults to 0)}

\item{two.stage}{indicator if the two-stage procedure for calculating the model weights should be used (defaults to TRUE)}
}
\value{
A list. The list contains the following named components:
\item{madr}{the model averaged double robust estimate}
\item{weight.ps}{a vector that contains the inclusion probability of each covariate in the propensity score model}
\item{weight.om}{a vector that contains the inclusion probability of each covariate in the outcome model}
}
\description{
This function estimates a model averaged double robust estimate.
}

