% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceFUN.R
\name{matern}
\alias{matern}
\title{## Matern Function
Useful for signals that approach a steady value but
have varied slopes at x=0, such as drawdown.}
\usage{
matern(x, params)
}
\description{
## Matern Function
Useful for signals that approach a steady value but
have varied slopes at x=0, such as drawdown.
}

