% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM.MA.enumerate.r
\name{OM.MA.enumerate}
\alias{OM.MA.enumerate}
\title{Enumerates all possible outcome models (linear terms only)}
\usage{
OM.MA.enumerate(Y, X, U, W = NULL)
}
\arguments{
\item{Y}{vector of the outcome}

\item{X}{vector of the treatment indicator (0/1)}

\item{U}{matrix of covariates to be considered for inclusion/exclusion}

\item{W}{matrix of covariates that will be included in all models (optional)}
}
\value{
A list. The listcontains the following named components:
\item{dict}{a list that contains the BIC, predicted values, and estimated treatment effect from each outcome model}
\item{out.table}{a matrix that contains the BIC and estimated treatment effect from each outcome model}
}
\description{
This function enumerates and fits all possible outcome models
}

