% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,MADproject,ANY-method}
\alias{plot,MADproject,ANY-method}
\title{Plot the MADproject object}
\usage{
\S4method{plot}{MADproject,ANY}(x)
}
\arguments{
\item{x}{The MADproject object}

\item{y}{not supported}

\item{...}{not supported}
}
\value{
NULL.
}
\description{
\code{plot.MADproject} plots ...
}

