% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.madr.enumerate.r
\name{print.madr.enumerate}
\alias{print.madr.enumerate}
\title{Print function for madr.enumerate class}
\usage{
\method{print}{madr.enumerate}(out)
}
\arguments{
\item{out}{madr.enumerate object}
}
\description{
This function prints results from madr.enumerate class
}

