% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS.MA.enumerate.r
\name{PS.MA.enumerate}
\alias{PS.MA.enumerate}
\title{Enumerates all possible propensity score models (linear terms only)}
\usage{
PS.MA.enumerate(X, U, W = NULL)
}
\arguments{
\item{X}{vector of the treatment indicator (0/1)}

\item{U}{matrix of covariates to be considered for inclusion/exclusion}

\item{W}{matrix of covariates that will be included in all models (optional)}
}
\value{
A list. The list contains the following named components:
\item{dict}{a list that contains the BIC and estimated propensity scores from propensity score models}
\item{out.table}{a matrix that contains the BIC from each propensity score model}
}
\description{
This function enumerates and fits all possible propensity score models
}

