% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pumping}
\alias{pumping}
\title{madr example data}
\format{A MADproject object with an observed time series, four samples,
and realizations for each sample.
\describe{
  \item{@observations}{a time series of drawdown, in meters}
  \item{@numSamples}{the value 4, [unitless]}
  \item{@realizations}{a list of length 4 with matrices
    containing an ensemble of 100 simulated drawdown time series,
    in meters}
  ...
}}
\source{
(temp)
}
\usage{
pumping
}
\description{
A dataset containing a MADproject object called 'pumping'
}
\keyword{datasets}

