% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceData.R
\name{reduceData}
\alias{reduceData}
\title{Apply the requested dimension reduction technique to the
inversion data for the given MADproject}
\usage{
reduceData(proj, method, params, ...)
}
\arguments{
\item{proj}{...}

\item{method}{...}

\item{params}{...}
}
\value{
proj
}
\description{
\code{reduceData} returns ...
}

