% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_convergence.R
\name{test_convergence}
\alias{test_convergence}
\title{Test the convergence of a MADproject object.}
\usage{
test_convergence(proj, data, ...)
}
\arguments{
\item{proj}{The MADproject object.}
}
\value{
NULL.
}
\description{
\code{test_convergence} returns a plot to help the user to visualize if
there are enough realizations in the project
}

