% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLikelihood.R
\name{calcLikelihood}
\alias{calcLikelihood}
\title{Calculate the likelihood for the samples in a MADproject object.}
\usage{
calcLikelihood(proj, dsubset, ...)
}
\arguments{
\item{proj}{The MADproject object with data read from the MAD# databases.}

\item{dsubset}{The subset of inversion data to use for the likelihood
calculations.}

\item{num_realz}{The number of realizations to use in the likelihood
calculation (defaults to all in the \code{realizations} slot)}

\item{samples}{A vector of sample IDs for which to calculate
likelihood values (defaults to all available in the
\code{realizations} slot)}
}
\value{
proj The updated MADproject object with a filled
\code{likelihood} slot.
}
\description{
\code{calcLikelihood} returns the likelihood values based on the
observation and realization data in MADproject or, optionally,
a subset thereof.
}
\details{
The likelihood calculation utilizes the \pkg{np} package for non-
parametric density estimation with all inversion data as dependent (i.e.
multivariate likelihood distributions are estimated).
}

