% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPosterior.R
\name{calcPosterior}
\alias{calcPosterior}
\title{Calculate the posterior for a MADproject object.}
\usage{
calcPosterior(proj, likes)
}
\arguments{
\item{proj}{The MADproject object with data read from the MAD# databases.}

\item{likes}{The likelihood values.}
}
\value{
NULL.
}
\description{
\code{calcPosterior} returns the posterior values ...
}

