% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madr.R
\docType{package}
\name{madr}
\alias{madr}
\alias{madr-package}
\title{madr: A package for post-processing of the Method of Anchored Distributions.}
\description{
The madr package provides three categories of important functions:
foo, bar and baz.
}
\section{foo functions}{

The foo functions ...
}

