% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madr.R
\docType{package}
\name{madr}
\alias{madr}
\alias{madr-package}
\title{madr: A package for post-processing of the Method of Anchored Distributions.}
\description{
The madr package provides functions for post-processing applications of
the Method of Anchored Distributions (MAD) such as viewing data, calculating
likelihood and posterior distributions,
testing convergence of likelihood values, and applying dimension reduction.
}

