% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.madr.mcmc.r
\name{print.madr.mcmc}
\alias{print.madr.mcmc}
\title{Print function for madr.mcmc class}
\usage{
\method{print}{madr.mcmc}(x, ...)
}
\arguments{
\item{x}{madr.mcmc object}

\item{...}{ignored}
}
\description{
This function prints results from madr.mcmc class
}

