% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.madr.enumerate.r
\name{print.summary.madr.enumerate}
\alias{print.summary.madr.enumerate}
\title{Print function for summary.madr.enumerate class}
\usage{
\method{print}{summary.madr.enumerate}(x, ...)
}
\arguments{
\item{x}{summary.madr.enumerate object}

\item{...}{ignored}
}
\description{
This function prints results from summary.madr.enumerate class
}

